Refactor.
As discussed in https://github.com/suluke/hfg-particles/commit/90126517d5c75e8ba74eddfb26471269372a46d7
